Add integration tests for swift-metadata-sync.
Added a framework and the first three tests for metadata sync. The
elasticsearch "cluster" is setup within the Docker container -- we may
want to move it out in the future.

Each will randomly generate an index, as well as any required
containers. Upon completion, the index and containers are removed.