Relax required verison of elasticsearch-py.
This allows us to use version 5.x.x or 6.x.x. As far as I can tell,
there are currently no breaking changes for us between the client
versions and we should work correctly with either one. In the future,
however, we may need to check the server version and import the
appropriate library (elasticsearch5 vs elasticsearch, or, presumably,
elasticsearch6).