Add Kafka handler
This commit adds a new kafka handler to the swift-metadata-sync process.
The purpose is to have a single process that is able to scan swift containers
for changes and send notification to either Elasticsearch and/or Kafka.

The Kafka pub/sub model is different in which it is not possible to query the
kafka endpoint to find out if a message has been prviously sent (as it is done
with other ContainerCrawler handlers). Thus we added the ability to record in
memcache once a message has been published. During the verification, memcached
is first checked to avoid publishing duplicate messages for the same container
row.