Update the internal library to support ES 7.x.
Elasticsearch has a breaking API change where document types are
deprecated and will be removed. Notably, in ES 7.x index and mapping
operations do not include document types. However, Elasticsearch 6.x
clusters may have document types (limited to one type per mapping) or
may not have a document type. As we were previously using "object" as
our document type, we have to be careful whether a new index has been
created without specifying a type (in which case it defaults to _doc
before being dropped). Or, there could be a case where this is an older
cluster with an index that has "object" for the document type.

The change also includes an update to the testing infrastructure:
1. packages Elasticsearch 7.4.0 server for testing, as well as 6.8.4
2. updated the logging to log to console