Fix mapping support for Elasticsearch 5.x.
While Elasticsearch 5.x is past its end of life, we may have users of
the older release. The patch makes sure we can continue to use an older
release with the eye of removing this support in the future.

The sample swift-metadata-sync.conf is updated for the current 7.4.0
version running in the container.

Unrelated fix:
A mapping may have been created with no properties. We should not assume
that properties is always set in the mappings response dictionary.